infer/requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
Pillow>=9.0.0

# API clients
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0

# Utilities
requests>=2.25.0
python-dotenv>=0.19.0
retry>=0.9.2 